requirements.txt
Django==1.3
Pygments==1.4
South==0.7.3
Whoosh==1.8.2
django-admin-tools==0.4.0
django-haystack==1.2.0
django-markitup==0.6.1
django-taggit==0.9.3
-e git+https://github.com/HonzaKral/django-threadedcomments.git#egg=django-threadedcomments
-e hg+https://bitbucket.org/ubernostrum/django-registration#egg=django-registration